Start FeedServer after seeding the database during feature test
In order for the genserver to initialize correctly, there needs to be
data available to use.

This ended up not working originally because I changed how the server
was started due to no longer passing in a dynamic name, which caused me
to pass in the list of options as the second arg, and not the third,
so the name was never being registred.