Install front end assets after elixir dependencies
This is because a couple of phoenix specific node packages are installed

Use a docker container with node and install front end assets in CI

Cache build artifacts for when the lockfile doesn't change

This is ripped directly from levelhq/level. I'm not certain of the
significance of the 'v3' number.

Update CircleCI config to use `test.all` mix task