Refactor(ish) to use a less naive supervision tree (and other stuff)
Wow, this was tough and I'm not sure if I'm perfectly satisfied, but I
am definitely happy enough.

The Old
===

Before, we started a single GenServer under the applications main
supervision tree, and was responsible for everything: fetching the blogs
from the database, fetching the feeds for each blog on a recurring
basis, merging them into a single feed, and also serving the built feed
to the Phoenix app when a request was made from the browser.

This created a massive single point of failure and was causing the site
to crash frequently.

The New
===

We now separate concerns and isolate different points of failure by
splitting up responsibilities into different processes.

The app will start a a ServerFarmSupervisor process which is responsible
for managing a ServerFarm, a FeedStore, and a couple "Cron" job processes.

ServerFarm
===

This is a DynamicSupervisor which is responsible for starting and
stopping "FeedServers"

FeedStore
===

This is a GenServer which is the primary liason between the phoenix
application and the whole "Server Farm & Friends" processes. This
includes serving the built feed, building the feed, and delegating
adding new feeds and removing deleted feeds to the ServerFarm.

FeedServer
===

This is a GenServer which does fetches a single blog and holds its
state. The ServerFarm will start a FeedServer for each blog, allowing us
to separate the HTTP requests into separate processes, so if one of them
fails, the rest are fine and the crashed process will be restarted by
the ServerFarm.

Cron
===

We also have two "Cron" jobs that are supervised by the
ServerFarmSupervisor.

Cron is a simple abstration of a recurring job that is based on a
GenServer. You just passed it a callback lambda and it will call it
recurringly according the configured timeout.

We have a "refresher" job, which will refresh the FeedStore by looping
through all the feed servers and "pushing" them to the FeedStore.

We also have a "write to disk" job, which is responsible for writing the
XML atom feed to a static file, so that rss readers can pick up on the
Planet blog.

Notes
===

Even after chiming in on a thread on the Elixir Forum
(https://elixirforum.com/t/what-are-your-thoughts-on-starting-processes-in-test-mode/15578/2)
and reading that you shouldn't start your application differently during
test env, I still decided that it was the best way to test this
application.