Implement Erlang releases for deployment.
This uses Distillery and Docker to build Erlang releases targeted at the
Ubuntu platform.

prod.exs changes are necessary for releases.

Includes a migrate release task that is used to run migrations on the
remote server since `mix` is not available.

Updates format configuration to format files in the /rel directory.