Add ability to configure the details of the Planet feed
This adds a `planets` table, which is used as a singleton row. The migration
will add the default row, and the user does not have any way to  delete
it, so it should be fine. (If they mess with the database manually,
¯\_(ツ)_/¯).

Also removed the `:type` field from the Feed struct because it was
unnecessary.